dev_requirements.txt
torch==1.4.0
torchvision==0.5.0
pytorch-lightning==0.7.1
azureml-sdk
# azureml-widgets